fixed issue with force sync
