+++ database/index.js
@@ -1,5 +1,6 @@
 const Sequelize = require('sequelize');
 const helper = require('./helper');
+const faker = require('faker');
 
 const sequelize = new Sequelize('menus', 'Joe', '', {
   host: 'localhost',
@@ -7,11 +8,45 @@ const sequelize = new Sequelize('menus', 'Joe', '', {
   dialect: 'postgres',
 });
 
+const menuNames = ['Breakfast', 'Lunch', 'Dinner'];
+
+const menuCategoryNames = ['Appetizers', 'Mains', 'Sides', 'Beverages'];
+
+const RestaurantMenuItems = sequelize.define('RestaurantMenuItems', {
+  restaurantId: Sequelize.INTEGER,
+  menuName: Sequelize.TEXT,
+  menuCategoryName: Sequelize.TEXT,
+  menuItemName: Sequelize.TEXT,
+  menuItemDescription: Sequelize.TEXT,
+  menuItemPrice: Sequelize.DECIMAL,
+});
+
+// creates restaurant data
+function createRestaurantData() {
+  for (let i = 1; i <= 200; i += 1) {
+    for (let j = 0; j < menuNames.length; j += 1) {
+      for (let k = 0; k < menuCategoryNames.length; k += 1) {
+        for (let l = 1; l <= helper.randomInt(5, 10); l += 1) {
+          RestaurantMenuItems.create({
+            restaurantId: i,
+            menuName: menuNames[j],
+            menuCategoryName: menuCategoryNames[k],
+            menuItemName: helper.capFirstLet(faker.lorem.words()),
+            menuItemDescription: faker.lorem.sentence().toLowerCase(),
+            menuItemPrice: helper.randomPrice(),
+          });
+        }
+      }
+    }
+  }
+}
 // overwrites current data and stores new
 sequelize
   .sync({
     force: true,
   })
   .then(() => {
-    helper.create();
+    createRestaurantData();
   });
+
+module.exports.create = createRestaurantData;
