+++ tests/seed.test.js
@@ -15,3 +15,4 @@ test('should capitalize the first letter of a string', () => {
   expect(string).toBe('Hello');
 });
 
+// test('should')
